requirements.txt
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
django-cors-headers==2.3.0
Pillow==5.2.0
raven==6.9.0
django-storages==1.6.6
django-push-notifications==1.6.0
psycopg2-binary==2.7.5
boto3==1.7.58
django-split-settings==0.3.0
